+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Contributed:
 
 Changes:
 
+- Use Promise resolve/reject to track state in typegen on-chain metadata retrievals
 - Basic support for markdown generation for different chains (only Kusama & Polkadot currently added)